+++ apps/probability/parameters_controller.h
@@ -8,7 +8,7 @@
 
 namespace Probability {
 
-class ParametersController : public Shared::FloatParameterController<float> {
+class ParametersController : public Shared::FloatParameterController<double> {
 public:
   ParametersController(Escher::Responder * parentResponder, Escher::InputEventHandlerDelegate * inputEventHandlerDelegate, Distribution * m_distribution, CalculationController * calculationController);
   const char * title() override;
@@ -24,8 +24,8 @@ private:
   Escher::HighlightCell * reusableParameterCell(int index, int type) override;
   int reusableParameterCellCount(int type) override;
   void buttonAction() override;
-  float parameterAtIndex(int index) override;
-  bool setParameterAtIndex(int parameterIndex, float f) override;
+  double parameterAtIndex(int index) override;
+  bool setParameterAtIndex(int parameterIndex, double f) override;
   bool textFieldDidFinishEditing(Escher::TextField * textField, const char * text, Ion::Events::Event event) override;
   class ContentView : public Escher::View {
   public:
